AuthUser middleware

+++ src/app/api/admin/delete-product/route.ts
@@ -1,31 +1,41 @@
 import connectToDB from '@/config/db';
 import { NextResponse } from 'next/server';
 import Product from '@/models/product';
+import AuthUser from '@/middleware/AuthUser';
 
 export const dynamic = 'force-dynamic';
 
 export async function DELETE(req: Request) {
   try {
     await connectToDB();
-    const id = new URL(req.url).searchParams.get('id');
+    const user = await AuthUser(req);
 
-    if (!id)
-      return NextResponse.json({
-        success: false,
-        message: 'Product ID is required!',
-      });
+    if (user?.role === 'admin') {
+      const id = new URL(req.url).searchParams.get('id');
 
-    const deletedProduct = await Product.findByIdAndDelete(id);
+      if (!id)
+        return NextResponse.json({
+          success: false,
+          message: 'Product ID is required!',
+        });
 
-    if (deletedProduct) {
-      return NextResponse.json({
-        success: true,
-        message: 'Product deleted successfully!',
-      });
+      const deletedProduct = await Product.findByIdAndDelete(id);
+
+      if (deletedProduct) {
+        return NextResponse.json({
+          success: true,
+          message: 'Product deleted successfully!',
+        });
+      } else {
+        return NextResponse.json({
+          success: false,
+          message: 'Failed to delete product. Please try again!',
+        });
+      }
     } else {
       return NextResponse.json({
         success: false,
-        message: 'Failed to delete product. Please try again!',
+        message: 'You are not authenticated!',
       });
     }
   } catch (error: any) {
